% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/world_administrative_boundaries.R
\docType{data}
\name{world_administrative_boundaries}
\alias{world_administrative_boundaries}
\title{world_administrative_boundaries}
\format{
A data frame with several rows and the following columns:
\describe{
\item{cca3}{Country code (3-letter).}
\item{geometry}{geographic boundaries}
}
}
\source{
Data obtained from the World Food Programme (UN agency) processed for use in this package.
}
\usage{
world_administrative_boundaries
}
\description{
A dataset containing information about world administrative boundaries.
}
\details{
This dataset includes country code and the geometries
}
\examples{
# Load the dataset and view the first few rows
data(world_administrative_boundaries)
head(world_administrative_boundaries)
}
\keyword{datasets}
